pyproject.toml
[project]
name = "tolmans-sowbug-playground"
version = "0.1.0"
description = "A research platform for studying psychological phenomena via agent-based simulation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "websockets>=14.0",
    "pyyaml>=6.0",
    "matplotlib>=3.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[project.scripts]
tolmans-sowbug-playground = "tolmans_sowbug_playground.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tolmans_sowbug_playground"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
